requirements.txt
# requirements.txt (core dependencies)
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
numpy==1.26.2
pydantic==2.5.2
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1

# requirements-dev.txt (development dependencies)
jupyterlab==4.0.9
ipykernel
notebook

# requirements-ml.txt (ML-specific dependencies)
scikit-learn
pandas==2.1.3
plotly==5.18.0
prometheus-api-client==0.5.4
ipywidgets==8.1.1

# Observability / Logging
loguru
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-requests

# Terminal Enhancements
rich
prometheus_client
python-dotenv
